Added new parse samplers, tested (non)uniform
The sampling functionality under LanguageModel was factored out into
its own file.  The sampler functions have been tested against
to determine whether they generate the desired distribution.

Uniform tests determine whether, when all link energies are equal, we
obtain a uniform distribution over labelled trees with n nodes.

Non-uniform tests determine whether, when link energies are not equal,
we get the target distribution, in which the probability of a tree is
proportional to the exp(sum of link energies).

Two parses, the one based on cycle-proof-rerooting and another based
on resampling nodes in contention have been found to be close to the
target distribution, but are slightly biased.  Hopefully this bias is
not important.