ParityTokenSampler working and sped up
I've finished the basic functions that are needed to resample tokens
given an initial list of tokens and a parse.  This is the companion
function to sampling parses.  One sampler generates possible parses
for a list of tokens, the other generates possible word substitutions
given a parse.

There were a couple points where I initially introduced steps that
required O(num_tokens * num_tokens * d) operation per sentence,
but it was possible to remove a factor of num_tokens from both,
and there's a good speedup, maybe cutting time by 60% for average
lengthed sentences.  It will be an even more important speedup for
padded datasets which have longer (padded) sentences.