Partially adapt ParityTokenResampler to take mask
I am in the process of adapting ParityTokenResampler.sample_tokens
to accept a masked input.  It seems to be correct, but there is
more testing to do on node_parity to be completely sure.

I am realizing that it is easier to have masked values appear as 0s
and keep ``mask`` as a separate input than it is to have a special
non-zero padding value such as -1.  This is because 0 behaves better
whenever a padded tensor is used for indexing, and because we can
always reassert the mask during calculations that could be affected
by the shape of the mask, and should always do so explicitly rather
assuming only that padded values won't throw errors when used as
indices but may return meaningless results.

This suggests that we should use 0 as padding everywhere.
An alternative would be to have a distinct padding token.  Then
we would use the id of this padding token in padded token tensors
while we would still use 0 as the padding value for any
sentence-position pointer tensors, like head_ptrs.

Decide this issue before moving on to updating
ParityTokenSampler.sample_parses, and update / test Dataset and
W2V model samplers and training accordingly.