ParityTokenResampler.sample_tokens accepts padding
ParityTokenResampler has been adapted to work with PaddingDataset.  I
adapted PaddingDataset to provide a padding mask as part of the batches
to make padding more explicit.

I am thinking about  including a <PAD> token in the vocabulary, to give
padding a dedicated value that isn't used for anything else in the
vocabulary.  This will make it easier to test for padding without
colliding with actual tokens.