Standardize train; make model viz
Encorporate dictionary into dataset and embeddings into model, to make
these more self-contained and to make it easier to factor out the
training loop for different models.

Create a model visualization that uses HTML and svg to show parse as
arcs between words.

Add temperature to EdgeModel.  As temperature rises, the probability
distribution smoothes out, becomes higher-entropy, and this can help to
escape perpetual loops in a walk-based or contention-based parser.

Enable ContentionParseSampler to take padding.