Make single-threaded, non-leaky, accept plain text
Add a "plain" parsing mode to make possible to parse the extracted,
sentence-split, and tokenized wikipedia dump.

It turns out that all of the effort to make this script multiprocessing
brings exactly no benefit.  So I made it single-threaded.

There was a memory leak in making dictionaries that I had to get rid of.
This is why the program would bog down and get killed.  It was due to
tokens being added to the dictionary keeping a reference to the entire
file contents alive (it's because tokens are just slices out of the original
text, and this is fixed by copying the desired data to new memory so
that the file contents can get garbage collected).