Add vocab limit to dictionary and dataset
Dictionary will still store it's full index of tokens, but for any
tokens whose index  is greater than vocab_limit it returns "<UNK>" or
the idx of "<UNK>".

Dictionary.__len__() returns either the length of the index of tokens or
vocab_limit if it is defined and smaller than the length of the index of
tkens.

Dataset knows how to calculate Px when vocabulary is limited.  It
credits the occurrences of tokens whose idx is greater than vocab_limit
as occurrences of "<UNK>", and then calculates Px on the basis of only
tokens under vocab_limit.