Get rid of other dataset representations
Eliminate LengthGroupedDataset, PaddedDataset.  Only use
PaddedDatasetParallel.  It would be nice though to have options to not
split batches into long and short sentences and to have the option to
guarantee that the sentence data iterates in the original order that it
appears on disk..

Fix bugs in PaddedDatasetParallel and improve organization of parsing code.

Fix bug in dictionary's application of vocab_limit that caused it to retain
vocab_limit + 1 tokens.

move train.view_edge to visualize.view_edge